feature/project-structure-implementation - http module moved under of src module
